+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.0] - 2026-07-05
+
 ### Changed
 
 - BBB: Send/receive all dCBWDataTransferLength data instead of STALLing endpoints (https://github.com/apohrebniak/usbd-storage/pull/26)
@@ -50,7 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[unreleased]: https://github.com/apohrebniak/usbd-storage/compare/v2.0.0...HEAD
+[unreleased]: https://github.com/apohrebniak/usbd-storage/compare/v3.0.0...HEAD
+[3.0.0]: https://github.com/apohrebniak/usbd-storage/releases/tag/v3.0.0
 [2.0.0]: https://github.com/apohrebniak/usbd-storage/releases/tag/v2.0.0
 [1.0.0]: https://github.com/apohrebniak/usbd-storage/releases/tag/v1.0.0
 [0.2.0]: https://github.com/apohrebniak/usbd-storage/releases/tag/v0.2.0